refactor(medication): move non-page modules out of pages directory
