+++ src/components/medication-active-ingredients/IngredientRow.tsx
