+++ src/components/medication-active-ingredients/types.ts
